% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_costs_label.R
\name{create_cost_label}
\alias{create_cost_label}
\title{Get live currencies}
\usage{
create_cost_label(name, currency)
}
\arguments{
\item{name}{name of cost variable}

\item{currency}{the three-letter code of the currency}
}
\value{
Returns the relevant cost label
}
\description{
Daily update with API from to EUR, from other currencies
}
