% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countries_map.R
\name{create_group_countries_map}
\alias{create_group_countries_map}
\title{Create a Group Countries Map}
\usage{
create_group_countries_map(
  country_maps,
  selected_country,
  selected_color = "#4C8187",
  unselected_color = "#E2DFDA",
  border_color = "#214E70"
)
}
\arguments{
\item{country_maps}{A named list of country polygons. Each element should be a spatial object that represents a country's shape.}

\item{selected_country}{A data frame or list containing the selected country. It should have a column or element named `country` with the country name.}

\item{selected_color}{A string representing the HEX code for the color of the selected country. Default is `"#4C8187"`.}

\item{unselected_color}{A string representing the HEX code for the color of the unselected countries. Default is `"#E2DFDA"`.}

\item{border_color}{A string representing the HEX code for the border color of the countries. Default is `"#214E70"`.}
}
\value{
A Leaflet map object with the countries' polygons added, highlighting the selected country.
}
\description{
This function creates a Leaflet map with polygons for each country, allowing a selected country to be highlighted with a different color.
}
\details{
This function adds polygons for each country to a Leaflet map. The selected country is highlighted with a different fill color, while other countries have a default fill color. The map's zoom, dragging, and scroll wheel zoom are disabled, and it is set to a fixed zoom level.
}
\examples{
\dontrun{
# Example usage:
country_maps <- list(USA = usa_polygon, Canada = canada_polygon, Mexico = mexico_polygon)
selected_country <- data.frame(country = "Canada")
map <- create_group_countries_map(country_maps, selected_country)
print(map)
}
}
