% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regions_map.R
\name{create_regions_map}
\alias{create_regions_map}
\title{Create a Regions Map}
\usage{
create_regions_map(
  selected_country,
  region_maps,
  selected_color = "#4C8187",
  unselected_color = "#E2DFDA",
  border_color = "#214E70"
)
}
\arguments{
\item{selected_country}{A string representing the name of the selected country.}

\item{region_maps}{A named list of region polygons for each country. Each element should be a spatial object that represents a country's regions.}

\item{selected_color}{A string representing the HEX code for the color of the selected region. Default is `"#4C8187"`.}

\item{unselected_color}{A string representing the HEX code for the color of the unselected regions. Default is `"#E2DFDA"`.}

\item{border_color}{A string representing the HEX code for the border color of the regions. Default is `"#214E70"`.}
}
\value{
A Leaflet map object with the regions' polygons added, highlighting the selected region.
}
\description{
This function creates a Leaflet map with polygons for each region within a selected country,
allowing a selected region to be highlighted with a different color.
}
\details{
This function adds polygons for each region within the selected country to a Leaflet map. The selected region is highlighted with a different fill color, while other regions have a default fill color. The map's zoom, dragging, and scroll wheel zoom are disabled, and it is set to a fixed zoom level.
}
\examples{
\dontrun{
# Example usage:
region_maps <- list(USA = usa_regions, Canada = canada_regions)
selected_country <- "Canada"
selected_region <- "Ontario"
map <- create_regions_map(selected_country, region_maps)
print(map)
}
}
