% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_live_currencies.R
\name{get_live_currency}
\alias{get_live_currency}
\title{Get live currencies}
\usage{
get_live_currency(api_key, base_currency, convert_to_api_currencies)
}
\arguments{
\item{api_key}{The temperature in degrees Fahrenheit}

\item{base_currency}{the currency it converts to}

\item{convert_to_api_currencies}{the other currencies it converts from}
}
\value{
The rates to calculate from currencies to the base currency
}
\description{
Daily update with API from to EUR, from other currencies
}
