% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_styles.R
\name{write_styles_css}
\alias{write_styles_css}
\title{Write or Append Custom styles.css to /www Directory}
\usage{
write_styles_css(primary, secondary)
}
\arguments{
\item{primary}{A string representing the HEX code for the primary color (e.g., "#FC1921").}

\item{secondary}{A string representing the HEX code for the secondary color (e.g., "#808284").}
}
\value{
A message indicating that the file has been written or appended to in the `/www` directory.
}
\description{
This function writes or appends to a `styles.css` file in the `/www` directory of the user's R project,
applying the user-defined primary and secondary colors for the headers.
}
\details{
This function checks if the `/www` directory exists in the current working directory.
If it does not exist, the function creates it and then writes the `styles.css` file with the specified colors.
If the `styles.css` file already exists, the function appends the new styles to it.
}
\examples{
\dontrun{
write_styles_css(primary = "#FC1921", secondary = "#808284")
}
}
